+++ api/configuration/v1alpha1/zz_generated_adopt_funcs.go
@@ -165,3 +165,13 @@ func (obj *KongSNI) GetAdoptOptions() *commonv1alpha1.AdoptOptions {
 func (obj *KongSNI) SetAdoptOptions(opts *commonv1alpha1.AdoptOptions) {
 	obj.Spec.Adopt = opts
 }
+
+// Get the options to adopt the resource from an existing resource.
+func (obj *KongDataPlaneClientCertificate) GetAdoptOptions() *commonv1alpha1.AdoptOptions {
+	return obj.Spec.Adopt
+}
+
+// Set the options to adopt the resource from an existing resource.
+func (obj *KongDataPlaneClientCertificate) SetAdoptOptions(opts *commonv1alpha1.AdoptOptions) {
+	obj.Spec.Adopt = opts
+}